feat: Add validations and a new way to export controller
